Add PlaywrightTests.cs and remove TitleShouldContainsPlaywright.cs
